Move renderer modules into core modules package
