+++ docs/research/life_update_strategy.md
@@ -2,11 +2,11 @@
 
 ## Problem Statement
 
-The Game of Life update in `heart.renderers.modules.life.state.LifeState` relied on a convolution-based neighbor count. This note captures the alternative padding-based neighbor count used to reduce per-frame overhead while keeping the same update rules.
+The Game of Life update in `heart.modules.life.state.LifeState` relied on a convolution-based neighbor count. This note captures the alternative padding-based neighbor count used to reduce per-frame overhead while keeping the same update rules.
 
 ## Materials
 
-- `src/heart/renderers/modules/life/state.py` for the update logic and neighbor counting.
+- `src/heart/modules/life/state.py` for the update logic and neighbor counting.
 - `src/heart/utilities/env.py` for configuration via environment variables.
 - `tests/modules/test_life_state.py` for parity checks between strategies.
 
@@ -17,6 +17,6 @@ The Game of Life update in `heart.renderers.modules.life.state.LifeState` relied
 
 ## Source Pointers
 
-- Neighbor counting logic: `heart.renderers.modules.life.state._count_neighbors_with_padding`.
-- Strategy selection: `heart.renderers.modules.life.state.LifeState._update_grid`.
+- Neighbor counting logic: `heart.modules.life.state._count_neighbors_with_padding`.
+- Strategy selection: `heart.modules.life.state.LifeState._update_grid`.
 - Configuration entry point: `heart.utilities.env.Configuration.life_update_strategy`.
