+++ src/heart/modules/water_cube/__init__.py
@@ -1,4 +1,4 @@
+from heart.modules.water_cube.provider import WaterCubeStateProvider
 from heart.peripheral.core.providers import container
-from heart.renderers.modules.water_cube.provider import WaterCubeStateProvider
 
 container[WaterCubeStateProvider] = WaterCubeStateProvider
