+++ src/heart/peripheral/providers/__init__.py
@@ -1,0 +1,1 @@
+"""Peripheral data providers for sensor and switch streams."""
