+++ src/heart/peripheral/providers/switch/__init__.py
@@ -1,4 +1,5 @@
 from heart.peripheral.core.providers import container
-from heart.renderers.modules.devices.switch.provider import MainSwitchProvider
+
+from .provider import MainSwitchProvider
 
 container[MainSwitchProvider] = MainSwitchProvider
